requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
python-dotenv==1.0.1
fastapi[standard]